backend/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.2
lxml>=4.9.3
cohere>=4.5.0
qdrant-client>=1.8.0
python-dotenv>=1.0.0
openai-agents>=0.1.0